+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm
 
+## 1.5.7
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.6
+
 ## 1.5.6
 
 ### Patch Changes