+++ packages/pqb/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pqb
 
+## 0.9.6
+
+### Patch Changes
+
+- Fix import timestamps issue
+
 ## 0.9.5
 
 ### Patch Changes